Update codebase to improve packaging and logging
In this update, the project configuration was adjusted to pack the solution instead of merely publishing it, which helps prepare it for distribution. The .csproj file was also updated to support both 'netstandard2.1' and 'net47' frameworks for wider compatibility. Changes were made in logging instances, replacing 'LogSource' with 'Logger' to ensure consistency throughout the codebase. PrivateAssets attributes were also included in the package references to avoid accidental exposure of internal packages. Code syntax has been cleaned up, particularly in FastBufferExtensions, to follow better code formatting practices.