Refactor BindToPreExistingObjectByBehaviour method and related code
The BindToPreExistingObjectByBehaviour method has been refactored to remove the 'InsertionPoint' parameter, which was used previously to limit where the method was applied. Now, the method is more flexible and can be applied without constraints. This change has also led to the removal of related code that handled the 'InsertionPoint' mechanism, including in the README, Plugin.cs, NetcodeValidator.cs and TextHandler.cs files, which makes the codebase simpler and cleaner.