Update documentation and logging for NetcodeValidator disposal
The README documentation for the instantiation and optional disposal of the `NetcodeValidator` was clarified. Users are now more explicitly advised that patches should be reverted only if necessary. The optional `OnDestroy()` method, which disposes the `NetcodeValidator`, was removed from the example. In addition, a log message that records when a plugin has unpatched all RPCs was added.